[fuse_ctrl, test] Move lib files into a single directory
