+++ src/integration/test_suites/libs/fuse_ctrl.h
@@ -21,7 +21,6 @@
 
 void lcc_initialization(void);
 void initialize_otp_controller(void);
-void reset_rtl(void);
 
 void sw_transition_req(uint32_t next_lc_state,
                         uint32_t token_127_96,
@@ -33,6 +32,5 @@ void sw_transition_req(uint32_t next_lc_state,
 void dai_wr(uint32_t addr, uint32_t wdata0, uint32_t wdata1, uint32_t granularity, uint32_t exp_mask);
 void dai_rd(uint32_t addr, uint32_t* rdata0, uint32_t* rdata1, uint32_t granularity, uint32_t exp_mask);
 void wait_dai_op_idle(uint32_t exp_mask);
-//void check_dai_status(void);
 
 #endif // FUSE_CTRL_H
