+++ src/integration/test_suites/smoke_test_fc_random_item_prov/smoke_test_fc_random_item_prov.c
@@ -9,6 +9,7 @@
 #include "soc_ifc.h"
 #include "fuse_ctrl_address_map.h"
 #include "caliptra_ss_lc_ctrl_address_map.h"
+#include "caliptra_ss_lib.h"
 #include "fuse_ctrl.h"
 #include "lc_ctrl.h"
 
@@ -48,7 +49,7 @@ void raw_to_testunlock0(){
     VPRINTF(LOW, "LC_CTRL: CALIPTRA_SS_LC_CTRL is in TEST_UNLOCK0 state!\n");
 }
 
-uint32_t calculate_digest(uint32_t partition_base_address) {
+void calculate_digest(uint32_t partition_base_address) {
     // Step 1: Check if DAI is idle
     wait_dai_op_idle(0);
 
@@ -112,6 +113,7 @@ void program_sw_manuf_partition(void) {
 
     // Step 4
     reset_rtl();
+    wait_dai_op_idle(0);
 
     // Step 5
     dai_rd(fuse_address, &read_data, NULL, 32, 0);
@@ -209,11 +211,11 @@ void main (void) {
     
     // Writing to Caliptra Boot GO register of MCI for CSS BootFSM to bring Caliptra out of reset 
     // This is just to see CSSBootFSM running correctly
-    lsu_write_32(SOC_MCI_REG_CALIPTRA_BOOT_GO, 1);
-    VPRINTF(LOW, "MCU: Writing MCI SOC_MCI_REG_CALIPTRA_BOOT_GO\n");
+    lsu_write_32(SOC_MCI_REG_CPTRA_BOOT_GO, 1);
+    VPRINTF(LOW, "MCU: Writing MCI SOC_MCI_REG_CPTRA_BOOT_GO\n");
 
-    uint32_t cptra_boot_go = lsu_read_32(SOC_MCI_REG_CALIPTRA_BOOT_GO);
-    VPRINTF(LOW, "MCU: Reading SOC_MCI_REG_CALIPTRA_BOOT_GO %x\n", cptra_boot_go);
+    uint32_t cptra_boot_go = lsu_read_32(SOC_MCI_REG_CPTRA_BOOT_GO);
+    VPRINTF(LOW, "MCU: Reading SOC_MCI_REG_CPTRA_BOOT_GO %x\n", cptra_boot_go);
     
     lcc_initialization();    
     raw_to_testunlock0();
