+++ src/integration/test_suites/smoke_test_fc_registers/smoke_test_fc_registers.c
@@ -9,6 +9,7 @@
 #include "soc_ifc.h"
 #include "fuse_ctrl_address_map.h"
 #include "caliptra_ss_lc_ctrl_address_map.h"
+#include "caliptra_ss_lib.h"
 #include "fuse_ctrl.h"
 #include "lc_ctrl.h"
 
@@ -94,11 +95,11 @@ void main (void) {
     
     // Writing to Caliptra Boot GO register of MCI for CSS BootFSM to bring Caliptra out of reset 
     // This is just to see CSSBootFSM running correctly
-    lsu_write_32(SOC_MCI_REG_CALIPTRA_BOOT_GO, 1);
-    VPRINTF(LOW, "MCU: Writing MCI SOC_MCI_REG_CALIPTRA_BOOT_GO\n");
+    lsu_write_32(SOC_MCI_REG_CPTRA_BOOT_GO, 1);
+    VPRINTF(LOW, "MCU: Writing MCI SOC_MCI_REG_CPTRA_BOOT_GO\n");
 
-    uint32_t cptra_boot_go = lsu_read_32(SOC_MCI_REG_CALIPTRA_BOOT_GO);
-    VPRINTF(LOW, "MCU: Reading SOC_MCI_REG_CALIPTRA_BOOT_GO %x\n", cptra_boot_go);
+    uint32_t cptra_boot_go = lsu_read_32(SOC_MCI_REG_CPTRA_BOOT_GO);
+    VPRINTF(LOW, "MCU: Reading SOC_MCI_REG_CPTRA_BOOT_GO %x\n", cptra_boot_go);
     
     lcc_initialization();    
     raw_to_testunlock0();
